APPPOCTOOL-27 Extend integration tests
