+++ pom.xml
@@ -36,7 +36,7 @@
     <spring-cloud-bom.version>2023.0.3</spring-cloud-bom.version>
     <openapi-tools.jackson-databind-nullable.version>0.2.6</openapi-tools.jackson-databind-nullable.version>
     <folio-spring-cql.version>8.1.2</folio-spring-cql.version>
-    <applications-poc-tools.version>1.5.6-SNAPSHOT</applications-poc-tools.version>
+    <applications-poc-tools.version>1.5.7-SNAPSHOT</applications-poc-tools.version>
     <folio-java-checkstyle.version>1.0.1</folio-java-checkstyle.version>
     <folio-flow-engine.version>1.1.1</folio-flow-engine.version>
