+++ src/test/java/org/folio/entitlement/it/EntitlementApplicationServiceIT.java
@@ -1,15 +1,7 @@
 package org.folio.entitlement.it;
 
-import static java.net.URLEncoder.encode;
-import static java.net.http.HttpRequest.BodyPublishers.ofString;
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static java.util.Objects.requireNonNull;
-import static org.folio.entitlement.support.TestConstants.HTTP_CLIENT_DUMMY_SSL;
-import static org.folio.entitlement.support.TestUtils.OBJECT_MAPPER;
 import static org.folio.test.security.TestJwtGenerator.generateJwtToken;
 import static org.hamcrest.Matchers.is;
-import static org.springframework.http.HttpMethod.POST;
-import static org.springframework.http.MediaType.APPLICATION_FORM_URLENCODED_VALUE;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.context.jdbc.Sql.ExecutionPhase.AFTER_TEST_METHOD;
 import static org.springframework.test.context.jdbc.Sql.ExecutionPhase.BEFORE_TEST_METHOD;
@@ -17,18 +9,10 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import java.net.URI;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse.BodyHandlers;
-import java.util.AbstractMap.SimpleImmutableEntry;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.stream.Collectors;
-import lombok.SneakyThrows;
-import org.apache.commons.lang3.StringUtils;
 import org.folio.common.utils.OkapiHeaders;
+import org.folio.entitlement.support.KeycloakTestClientConfiguration;
+import org.folio.entitlement.support.KeycloakTestClientConfiguration.KeycloakTestClient;
 import org.folio.entitlement.support.base.BaseIntegrationTest;
-import org.folio.security.integration.keycloak.configuration.properties.KeycloakProperties;
 import org.folio.test.extensions.EnableKeycloakSecurity;
 import org.folio.test.extensions.EnableKeycloakTlsMode;
 import org.folio.test.extensions.KeycloakRealms;
@@ -36,12 +20,14 @@ import org.folio.test.extensions.WireMockStub;
 import org.folio.test.types.IntegrationTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
 
-@EnableKeycloakTlsMode
 @IntegrationTest
+@EnableKeycloakTlsMode
 @EnableKeycloakSecurity
 @KeycloakRealms("/keycloak/test-realm.json")
+@Import(KeycloakTestClientConfiguration.class)
 @Sql(executionPhase = BEFORE_TEST_METHOD, scripts = "classpath:/sql/folio-entitlement.sql")
 @Sql(executionPhase = AFTER_TEST_METHOD, scripts = "classpath:/sql/truncate-tables.sql")
 class EntitlementApplicationServiceIT extends BaseIntegrationTest {
@@ -51,11 +37,8 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
   private static final String APPLICATION_VERSION = "2.0.0";
   private static final String TEST_TENANT = "test";
   private static final String TEST_TENANT2 = "test2";
-  private static final Map<String, Entry<String, String>> REALM_CREDENTIALS_MAP = Map.of(
-    TEST_TENANT, new SimpleImmutableEntry<>("test-login-application", "test-login-application-secret"),
-    TEST_TENANT2, new SimpleImmutableEntry<>("test2-login-application", "test2-login-application-secret"));
 
-  @Autowired private KeycloakProperties keycloakProperties;
+  @Autowired private KeycloakTestClient keycloakTestClient;
 
   @Test
   @WireMockStub(scripts = {
@@ -66,7 +49,7 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
     mockMvc.perform(get("/entitlements/{tenantName}/applications", TEST_TENANT)
         .contentType(APPLICATION_JSON)
         .header(OkapiHeaders.TENANT, TEST_TENANT)
-        .header(OkapiHeaders.TOKEN, generateAccessToken(TEST_TENANT)))
+        .header(OkapiHeaders.TOKEN, keycloakTestClient.generateAccessToken(TEST_TENANT)))
       .andExpect(status().isOk())
       .andExpect(jsonPath("$.totalRecords", is(1)))
       .andExpect(jsonPath("$.applicationDescriptors[0].id", is(APPLICATION_ID)))
@@ -80,7 +63,7 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
     mockMvc.perform(get("/entitlements/{tenantName}/applications", TEST_TENANT2)
         .contentType(APPLICATION_JSON)
         .header(OkapiHeaders.TENANT, TEST_TENANT)
-        .header(OkapiHeaders.TOKEN, generateAccessToken(TEST_TENANT)))
+        .header(OkapiHeaders.TOKEN, keycloakTestClient.generateAccessToken(TEST_TENANT)))
       .andExpect(status().isOk());
   }
 
@@ -90,7 +73,7 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
     mockMvc.perform(get("/entitlements/{tenantName}/applications", TEST_TENANT)
         .contentType(APPLICATION_JSON)
         .header(OkapiHeaders.TENANT, TEST_TENANT)
-        .header(OkapiHeaders.TOKEN, generateAccessToken(TEST_TENANT2)))
+        .header(OkapiHeaders.TOKEN, keycloakTestClient.generateAccessToken(TEST_TENANT2)))
       .andExpect(status().isForbidden())
       .andExpect(jsonPath("$.total_records", is(1)))
       .andExpect(jsonPath("$.errors[0].message", is("X-Okapi-Tenant header is not the same as resolved tenant")))
@@ -103,7 +86,7 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
     mockMvc.perform(get("/entitlements/{tenantName}/applications", TEST_TENANT)
         .contentType(APPLICATION_JSON)
         .header(OkapiHeaders.TENANT, "another-tenant")
-        .header(OkapiHeaders.TOKEN, generateAccessToken(TEST_TENANT)))
+        .header(OkapiHeaders.TOKEN, keycloakTestClient.generateAccessToken(TEST_TENANT)))
       .andExpect(status().isForbidden())
       .andExpect(jsonPath("$.total_records", is(1)))
       .andExpect(jsonPath("$.errors[0].message", is("X-Okapi-Tenant header is not the same as resolved tenant")))
@@ -123,32 +106,5 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
       .andExpect(jsonPath("$.errors[0].message", is("Failed to validate a token")))
       .andExpect(jsonPath("$.errors[0].type", is("NotAuthorizedException")))
       .andExpect(jsonPath("$.errors[0].code", is("auth_error")));
-  }
-
-  @SneakyThrows
-  private String generateAccessToken(String tenant) {
-    var credentials = requireNonNull(REALM_CREDENTIALS_MAP.get(tenant));
-    var tokenRequestBody = Map.of(
-      "client_id", credentials.getKey(),
-      "client_secret", credentials.getValue(),
-      "grant_type", "client_credentials");
-
-    var keycloakBaseUrl = StringUtils.removeEnd(keycloakProperties.getUrl(), "/");
-    var uri = URI.create(String.format("%s/realms/%s/protocol/openid-connect/token", keycloakBaseUrl, tenant));
-    var request = HttpRequest.newBuilder(uri)
-      .method(POST.name(), ofString(toFormUrlencodedValue(tokenRequestBody), UTF_8))
-      .header("Content-Type", APPLICATION_FORM_URLENCODED_VALUE)
-      .build();
-
-    var response = HTTP_CLIENT_DUMMY_SSL.send(request, BodyHandlers.ofString(UTF_8));
-    var keycloakTokenJson = OBJECT_MAPPER.readTree(response.body());
-    return keycloakTokenJson.path("access_token").asText();
-  }
-
-  private static String toFormUrlencodedValue(Map<String, String> params) {
-    return params.entrySet()
-      .stream()
-      .map(entry -> String.format("%s=%s", encode(entry.getKey(), UTF_8), encode(entry.getValue(), UTF_8)))
-      .collect(Collectors.joining("&"));
   }
 }
