+++ src/test/resources/wiremock/mgr-applications/folio-app1/get-by-ids-query-full.json
@@ -5,6 +5,9 @@
     "headers": {
       "content-type": {
         "equalTo": "application/json"
+      },
+      "x-okapi-token": {
+        "matches": ".+"
       }
     },
     "queryParameters": {
