+++ src/test/resources/wiremock/mgr-applications/folio-app1/get-discovery.json
@@ -10,6 +10,9 @@
     "headers": {
       "content-type": {
         "equalTo": "application/json"
+      },
+      "x-okapi-token": {
+        "matches": ".+"
       }
     }
   },
