+++ src/test/resources/wiremock/mgr-tenants/test/get-404.json
@@ -7,7 +7,7 @@
         "equalTo": "application/json"
       },
       "x-okapi-token": {
-        "equalTo": "X-Okapi-Token test value"
+        "matches": ".+"
       }
     }
   },
