+++ src/test/resources/wiremock/mgr-tenants/test/get-query-by-name.json
@@ -5,6 +5,9 @@
     "headers": {
       "content-type": {
         "equalTo": "application/json"
+      },
+      "x-okapi-token": {
+        "matches": ".+"
       }
     },
     "queryParameters": {
