+++ src/test/resources/wiremock/mod-authtoken/verify-token-any.json
@@ -4,7 +4,7 @@
     "urlPathPattern": ".*",
     "headers": {
       "X-Okapi-Token": {
-        "equalTo": "X-Okapi-Token test value"
+        "matches": ".+"
       },
       "X-Okapi-Tenant": {
         "equalTo": "supertenant"
